Updated the file
Managed to clean up the table so that it used more meaningful values, got rid of columns that weren't seen to really correlate to the price (in hindsight, should have use the individual correlation messages to do this rather than just manually doing it based on 'intuition'). Also wrote a knn.reg() 'calculation'. Haven't been able to do much meaningful work with the data yet however (in that the predictions/error rates are something that I haven't even reached yet). I am VERY behind on this. To be honest, I didn't really understand what I was doing at all yesterday and I understand now but it feels like the level of understanding I have now is what everyone else had yesterday and that I won't get this done.