Uploaded knn_regression
Have written a script with knn regression in order to get a prediction of sale price (managed to get it within a 5% error -- not sure how this happened so will look more at data tomorrow). Had to clean up the data (adding in dummy variables) prior to the regression itself.